backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
openai-whisper==20231117
lxml==4.9.3
torch==2.2.0
torchaudio==2.2.0
